Made some methods and classes more cleaned and suspended
