+++ .idea/workspace.xml
@@ -29,16 +29,20 @@
     <select />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="ce137a6b-3ca0-496a-a1d1-8eccf2f23434" name="Default Changelist" comment="added realm database like local store for data">
+    <list default="true" id="ce137a6b-3ca0-496a-a1d1-8eccf2f23434" name="Default Changelist" comment="1) Added SwipeRefreshLayout for refresh database&#10;2) Koin injection for UserListViewModel">
+      <change afterPath="$PROJECT_DIR$/data/src/main/java/com/wainow/data/db/DBHelper.kt" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.gradle/6.5/executionHistory/executionHistory.bin" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/6.5/executionHistory/executionHistory.bin" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.gradle/6.5/executionHistory/executionHistory.lock" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/6.5/executionHistory/executionHistory.lock" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.gradle/6.5/fileHashes/fileHashes.bin" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/6.5/fileHashes/fileHashes.bin" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.gradle/6.5/fileHashes/fileHashes.lock" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/6.5/fileHashes/fileHashes.lock" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.gradle/buildOutputCleanup/buildOutputCleanup.lock" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/buildOutputCleanup/buildOutputCleanup.lock" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.gradle/buildOutputCleanup/outputFiles.bin" beforeDir="false" afterPath="$PROJECT_DIR$/.gradle/buildOutputCleanup/outputFiles.bin" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/presentation/src/main/java/com/wainow/friendzone/view/activity/MainActivity.kt" beforeDir="false" afterPath="$PROJECT_DIR$/presentation/src/main/java/com/wainow/friendzone/view/activity/MainActivity.kt" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/presentation/src/main/java/com/wainow/friendzone/view/fragment/UserListFragment.kt" beforeDir="false" afterPath="$PROJECT_DIR$/presentation/src/main/java/com/wainow/friendzone/view/fragment/UserListFragment.kt" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/presentation/src/main/res/layout/user_list_activity.xml" beforeDir="false" afterPath="$PROJECT_DIR$/presentation/src/main/res/layout/user_list_activity.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/data/src/main/java/com/wainow/data/repository/UserRepositoryImpl.kt" beforeDir="false" afterPath="$PROJECT_DIR$/data/src/main/java/com/wainow/data/repository/UserRepositoryImpl.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/domain/src/main/java/com/wainow/domain/interactor/UserInteractor.kt" beforeDir="false" afterPath="$PROJECT_DIR$/domain/src/main/java/com/wainow/domain/interactor/UserInteractor.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/domain/src/main/java/com/wainow/domain/interactor/UserInteractorImpl.kt" beforeDir="false" afterPath="$PROJECT_DIR$/domain/src/main/java/com/wainow/domain/interactor/UserInteractorImpl.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/domain/src/main/java/com/wainow/domain/repository/UserRepository.kt" beforeDir="false" afterPath="$PROJECT_DIR$/domain/src/main/java/com/wainow/domain/repository/UserRepository.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/presentation/src/main/java/com/wainow/friendzone/view/fragment/UserListViewModel.kt" beforeDir="false" afterPath="$PROJECT_DIR$/presentation/src/main/java/com/wainow/friendzone/view/fragment/UserListViewModel.kt" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -231,7 +235,14 @@
       <option name="project" value="LOCAL" />
       <updated>1623445363186</updated>
     </task>
-    <option name="localTasksCounter" value="9" />
+    <task id="LOCAL-00009" summary="1) Added SwipeRefreshLayout for refresh database&#10;2) Koin injection for UserListViewModel">
+      <created>1623481404679</created>
+      <option name="number" value="00009" />
+      <option name="presentableId" value="LOCAL-00009" />
+      <option name="project" value="LOCAL" />
+      <updated>1623481404680</updated>
+    </task>
+    <option name="localTasksCounter" value="10" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -255,7 +266,8 @@
     <MESSAGE value="added implicit intents" />
     <MESSAGE value="1) added scroll view for small screens&#10;2) fixed a bug with toolbar arrows" />
     <MESSAGE value="added realm database like local store for data" />
-    <option name="LAST_COMMIT_MESSAGE" value="added realm database like local store for data" />
+    <MESSAGE value="1) Added SwipeRefreshLayout for refresh database&#10;2) Koin injection for UserListViewModel" />
+    <option name="LAST_COMMIT_MESSAGE" value="1) Added SwipeRefreshLayout for refresh database&#10;2) Koin injection for UserListViewModel" />
   </component>
   <component name="WindowStateProjectService">
     <state x="245" y="186" key="#PSD" timestamp="1622750323203">
@@ -286,22 +298,22 @@
       <screen x="0" y="0" width="1440" height="900" />
     </state>
     <state x="538" y="214" key="AndroidCreateXmlResourceDialog/0.0.1440.900@0.0.1440.900" timestamp="1623345183422" />
-    <state width="1396" height="245" key="GridCell.Tab.0.bottom" timestamp="1623481219850">
+    <state width="1396" height="245" key="GridCell.Tab.0.bottom" timestamp="1623483460491">
       <screen x="0" y="0" width="1440" height="900" />
     </state>
-    <state width="1396" height="245" key="GridCell.Tab.0.bottom/0.0.1440.900@0.0.1440.900" timestamp="1623481219850" />
-    <state width="1396" height="245" key="GridCell.Tab.0.center" timestamp="1623481219849">
+    <state width="1396" height="245" key="GridCell.Tab.0.bottom/0.0.1440.900@0.0.1440.900" timestamp="1623483460491" />
+    <state width="1396" height="245" key="GridCell.Tab.0.center" timestamp="1623483460489">
       <screen x="0" y="0" width="1440" height="900" />
     </state>
-    <state width="1396" height="245" key="GridCell.Tab.0.center/0.0.1440.900@0.0.1440.900" timestamp="1623481219849" />
-    <state width="1396" height="245" key="GridCell.Tab.0.left" timestamp="1623481219848">
+    <state width="1396" height="245" key="GridCell.Tab.0.center/0.0.1440.900@0.0.1440.900" timestamp="1623483460489" />
+    <state width="1396" height="245" key="GridCell.Tab.0.left" timestamp="1623483460487">
       <screen x="0" y="0" width="1440" height="900" />
     </state>
-    <state width="1396" height="245" key="GridCell.Tab.0.left/0.0.1440.900@0.0.1440.900" timestamp="1623481219848" />
-    <state width="1396" height="245" key="GridCell.Tab.0.right" timestamp="1623481219850">
+    <state width="1396" height="245" key="GridCell.Tab.0.left/0.0.1440.900@0.0.1440.900" timestamp="1623483460487" />
+    <state width="1396" height="245" key="GridCell.Tab.0.right" timestamp="1623483460490">
       <screen x="0" y="0" width="1440" height="900" />
     </state>
-    <state width="1396" height="245" key="GridCell.Tab.0.right/0.0.1440.900@0.0.1440.900" timestamp="1623481219850" />
+    <state width="1396" height="245" key="GridCell.Tab.0.right/0.0.1440.900@0.0.1440.900" timestamp="1623483460490" />
     <state width="1396" height="245" key="GridCell.Tab.1.bottom" timestamp="1623442222221">
       <screen x="0" y="0" width="1440" height="900" />
     </state>
