+++ domain/src/main/java/com/wainow/domain/interactor/UserInteractorImpl.kt
@@ -5,7 +5,7 @@ import com.wainow.domain.repository.UserRepository
 
 class UserInteractorImpl(private val userRepository: UserRepository): UserInteractor {
     override suspend fun getUsers() = userRepository.getUsers()
-    override fun saveUsers(users: List<User>) { userRepository.saveUsers(users) }
-    override fun getLocalUsers(): List<User> = userRepository.getLocalUsers()
-    override fun deleteUsers() { userRepository.deleteUsers() }
+    override suspend fun saveUsers(users: List<User>) { userRepository.saveUsers(users) }
+    override suspend fun getLocalUsers(): List<User> = userRepository.getLocalUsers()
+    override suspend fun deleteUsers() { userRepository.deleteUsers() }
 }
