+++ domain/src/main/java/com/wainow/domain/repository/UserRepository.kt
@@ -4,7 +4,7 @@ import com.wainow.domain.entity.User
 
 interface UserRepository {
     suspend fun getUsers(): List<User>
-    fun getLocalUsers(): List<User>
-    fun saveUsers(users: List<User>)
-    fun deleteUsers()
+    suspend fun getLocalUsers(): List<User>
+    suspend fun saveUsers(users: List<User>)
+    suspend fun deleteUsers()
 }
